refactor: move some retort code to render module
